+++ tests/testthat/test_checkCharacter.R
@@ -15,10 +15,14 @@ test_that("checkCharacter", {
 
   expect_true(testCharacter("a", min.chars = 1))
   expect_false(testCharacter("a", min.chars = 2))
+
   # treat NA_character_ as zero-length string
   expect_true(testCharacter(NA_character_, min.chars = 0))
-  expect_false(testCharacter(NA_character_, min.chars = 1))
-  expect_false(testCharacter(NA, min.chars = 1))
+  expect_true(testCharacter(NA_character_, min.chars = 1))
+  expect_false(testCharacter(NA_character_, min.chars = 1, any.missing = FALSE))
+  expect_false(testCharacter(c("", NA_character_), min.chars = 1))
+  expect_true(testCharacter(NA, min.chars = 1))
+  expect_true(testCharacter(character(0), min.chars = 1))
 
   x = c("abba", "baab")
   expect_true(testCharacter(x, pattern="a"))